SLIDESCLOUDDOC-50, SLIDESCLOUDDOC-47 The article "Delete Paragraphs" is added
The chapter "Working with Paragraphs" is added.
The articles about paragraphs are moved to the chapter "Working with Paragraphs".
The articles "Remove a Paragraph" and "Remove a Paragraph Range" are deleted.
The article "Delete Paragraphs" is added.
SLIDESCLOUDDOC-50: Code examples in cURL, C#, Java, PHP, Ruby, Python, Node.js, C++, and Perl are added for the DeleteParagraph method.
SLIDESCLOUDDOC-47: Code examples in cURL, C#, Java, PHP, Ruby, Python, Node.js, C++, and Perl are added for the DeleteParagraphs method.