SLIDESCLOUDDOC-179 The article about deleting comments is updated
The articles "Delete Comments" and "Delete Slide Comments" are merged to "Delete Comments".
Information and code examples for the DeleteComments are updated, the code example in C++ is added.
Information about the DeleteCommentsOnline method is added.
Information and code examples for the DeleteSlideComments are updated, the code examples in Node.js and C++ are added.