SLIDESCLOUDDOC-196 The article "Update an Effect in a Main Sequence" is added is added
The article "Update a Main Sequence" is renamed to "Add an Effect to a Main Sequence". Minor corrections are performed.
The article "Update an Effect in a Main Sequence":
Introduction, API information, and code examples in cURL, C#, Java, PHP, Ruby, Python, Node.js, C++, and Perl for the UpdateAnimationEffect method are added.