SLIDESCLOUDDOC-227 The article "Working with Slide Show Properties" is updated
Keywords are added and the introduction is updated.
The API information and code examples in cURL, C#, Java, PHP, Ruby, Python, Node.js, and Go for the GetSlideShowProperties method are updated.
Code examples in C++ and Perl for the GetSlideShowProperties method are added.
A picture is added.
The API information and code examples in cURL, C#, Java, PHP, Ruby, Python, Node.js, and Go for the SetSlideShowProperties method are updated.
Code examples in C++ and Perl for the SetSlideShowProperties method are added.