SLIDESCLOUDDOC-228 The article "Working with View Properties" is updated
Keywords are added and the introduction is updated.
The API information and code examples in cURL, C#, Java, PHP, Ruby, Python, Node.js, and Go for the GetViewProperties method are updated.
Code examples in C++ and Perl for the GetViewProperties method are added.
The API information and code examples in cURL, C#, Java, PHP, Ruby, Python, Node.js, and Go for the SetViewProperties method are updated.
Code examples in C++ and Perl for the SetViewProperties method are added.