SLIDESCLOUDDOC-237,238 The articles "Get Paragraph" and "Get Paragraphs" are updated
The article "Get Paragraphs" is updated:
The introduction, API information, example description, and code examples in cURL, C#, Java, PHP, Ruby, Python, Node.js, Perl, and Go are updated.
Pictures and code example in C++ are added.

The article "Get a Paragraph" is updated:
The API information, example description, and code examples in cURL, C#, Java, PHP, Ruby, Python, Node.js, Perl, and Go are updated.
A code example in C++ is added.
The article is merged to"Get Paragraphs".